requirements.txt
certifi==2016.2.28
dj-database-url==0.4.1
Django==1.11.1
django-extensions==1.5.6
django-filter==0.11.0
djangorestframework==3.2.3
gunicorn==19.6.0
Markdown==2.6.2
newrelic==2.54.0.41
psycopg2==2.6.2
pytz==2017.2
six==1.9.0
whitenoise==3.2
